refactor(kmp): 完成 CMP 跨端架构迁移并精简工程
- 将 Home、Shorts、Detail 页面迁移为 Android/iOS 共用的 CMP Screen
- 抽取共享 StateHolder、Presenter、分页状态和领域仓库接口
- 为 Android 接入 Coil、Media3、Navigation 3 和生命周期适配
- 为 iOS 接入 Foundation、UIKit、AVPlayer 和 CMP 宿主
- 修复 iOS 启动尺寸、图片域名和模拟器架构问题
- 删除 Lib_Network、legacy network、旧 SwiftUI 页面及无用模板资源
- 优化 Gradle 模块依赖、网络安全和播放器生命周期
- 增加跨端回归测试、架构守卫及 KMP/CMP 新人教程

BREAKING CHANGE: 移除旧网络模块和旧 Android/SwiftUI 页面实现

+++ core_data/src/iosMain/kotlin/com/kotlinmvvm/core/data/eyepetizer/EyepetizerPayloadParser.ios.kt
@@ -90,7 +90,7 @@ private fun Map<*, *>.intValue(key: String): Int? {
         is Long -> value.toInt()
         is Double -> value.toInt()
         is Float -> value.toInt()
-        is NSNumber -> value.intValue.toInt()
+        is NSNumber -> value.intValue
         else -> null
     }
 }
